feat(clothes): create DTOs for clothes, variants, and allowed images
